Fix StrictHostKeyChecking-related failure
For ssh-through git connections, the git pull fails because the host key will not be automatically added to known_hosts.

This should fix that by allowing the key to be put into the known_hosts.

Note that if the remote server ssh key changes, the git operation will fail. That seems a reasonably safe security fallback.