Re-enable ssh add-on's network ssh access.
PR #1117 disabled SSH network access for the SSH server add-on.  

Users who previous had the "SSH Server" official add-on installed, will be surprised when they update to version 8.4 that by default ssh access from the network is no longer available.  

PR #1117 has effectively changed this add-on from something that enabled ssh access over the network into something that only provides access through a web GUI terminal.

This PR simply makes ssh enabled be default restoring the previous behavior to avoid frustrating existing users. 

If the intent is to have SSH network access disabled by default. it should be better communicated as a breaking change in a future PR.  There are a number of ways this could be done.