letsencrypt: Adds cron based daily renewal check and restart on success.
This adds some a basic crond based daily check for the certificate
renewal.  It will run the check as a defined 24h:min based time that
the user can define.  As well it allows for the restart of the core
and defined addons by slug on a successful renewal and deployment of
the cert.