Updated readme to include invalid Cert files
While trying to use this add-on I noticed that even if the certfile setting is false, if blank /ssl/privkey.pem and /ssl/fullchain.pem files existed, it would not work. I needed to delete these files. I don't necessarily view this as a bug, so I figured adding it to the documentation covers it.