ssh: Mention port setting in step by step instructions
When trying to use the ssh add-on, I was very confused by that "null"
value in the plugin configuration, and why I could not access my
home-assistant instance by ssh, ... It was not clear to me that the
initial mention of port 22 was about configuration in the add-on
configuration or some configuration to do on your home network.

Moving the port description to the step by step instructions for
configuring the add-on should clear that up.