Add/Fix Azure section
I went through this and couldn't get this to work without looking at error messages
first problem was it didn't say that i need the azure_config and what that did. I corrected that (row 165)
second problem was a style issue? or me trying to fix this based on experience.  I'd have tweaked more but ran out of time.