Use version of node-zwave-js with xstate pinned
That version is a test build, which is basically 10.2.0, but with xstate pinned to 4.29.0 - the version node-zwave-js had in its own lockfile. This seems to be the most obvious difference in dependencies between the addons.

With this change, the core addon will use an older version of the dependency than the community addon (4.32.1), but one that was previously in use and did not leak.