Fix slow startup of dnsmasq on HAOS 10
By default dnsmasq seems to close all potentially open file descriptors.
This has previously brought up on the mailing list [1], but it seems
that the fix did not really fix the issue (maybe an Alpine issue).

Previously the limit of open file descriptors was set to 1048576 for all
add-ons (see [2]), which depending on the system takes up to 1s for
dnsmasq to start. With the new, much higher limit it takes minutes to
start dnsmasq.

This changes the limit to 1024, which speeds up dnsmasq startup in both
cases to be well below 1s.

[1] https://lists.thekelleys.org.uk/pipermail/dnsmasq-discuss/2020q1/013821.html
[2] https://developers.home-assistant.io/blog/2023/04/13/new_limits_for_add_ons