openthread_border_router: update OTBR and universal-silabs-flasher (#3134)
* openthread_border_router: update OTBR and universal-silabs-flasher

Use the latest OTBR version which allows to use the DELETE method to
the /node endpoint to erase all persistent information of the node.

* Explicitly install the OpenBSD version of netcat

In bullseye the netcat package depends on netcat-openbsd, so this
doesn't change anything in practice. However, in newer Debian releases
the package is a virtual package, hence a concrete package must be used.

* openthread_border_router: avoid API breakage

Backport the dataset delete API to avoid breaking the API. Adjust
version to indicate no API breakage.

* openthread_border_router: use hassio hostname

Use hassio hostname in discovery information. With that the add-on
should also work even when not exposing the ports.