matter_server: Add beta flag (#3282)
* matter_server: Add beta flag

Add a new flag which allows to install the latest pre-release of the
Python Matter Server. This is useful for testing new (pre-)releases
before bumping the add-on.

Previously we tried different approaches like a separate beta version of
the add-on. This has the problem that it has it's separate data store
(means no more devices in case the non-beta add-on has been used
previously) and it gets chaotic as the Matter integration is still
trying to start the non-beta version of the add-on.

With this attempt we can reuse the same add-on for testing. The downside
is once new OS dependencies are required, those won't be added. In that
case a regular add-on version bump would be necessary first.

* Update matter_server/config.yaml

Co-authored-by: Franck Nijhof <frenck@frenck.nl>

---------

Co-authored-by: Franck Nijhof <frenck@frenck.nl>