silabs-multiprotocol: Make the OTBR infra netif configurable
The OpenThread Border Router agent currently only supports routing
between a single "infrastructure" network interface and a single Thread
network interface per agent process.  (Multiple "infrastructure"
interfaces can be specified, but this is for fail-over purposes and only
one will be used at a time.)

Thread device commissioning in HA currently requires an Android device
which can communicate both with the Thread device via Bluetooth and with
the OpenThread Border Router "infrastructure" interface via IPv6 with
Router Advertisements and mDNS.

Prior to this commit, the first HA host interface with an associated
default route was always used as the OTBR "infrastructure" interface.
This caused problems if the HA host has multiple interfaces and the
Android device needed for commissioning is on a different network than
the automatically-selected interface.

To address that, this commit permits the user to configure the
"infrastructure" interface.