openthread_border_router: Service improvements and OTBR bump
Make sure otbr-agent shutdowns lead to actual container shutdown.
Currently a rather crude workaround is required to get this done. This
seems to work with two test cases:
- Non-working/non-responding radio (e.g. serial port with nothing
  attached)
- OTBR REST API port in use

Also make sure that the mDNS service is a dependency of the OTBR Agent
to ensure that mDNS is shutdown after the OTBR.

Also bump the OTBR and SiLabs Universal Flasher to the latest versions.