mariadb: Increase timeout for MariaDB shutdown (#3570)
* mariadb: Increase timeout for MariaDB shutdown

Currently, when MariaDB takes longer than 20s to shutdown, Docker will
simply kill it. This can be problematic on upgrade, as the new MariaDB
version cannot handle a not cleanly shutdown MariaDB (see #3566).

Increase the timeout of the container shutdown to 60s. Also define a
per-service kill timeout for mariadb-core. This is largely optional, and
essentially leads to the same outcome if MariaDB takes even longer (it
will get killed, one way or another). But by defining this as part of
the s6 service makes sure that the add-on log itself shows that MariaDB
had to be killed, and makes sure the rest of the s6 service tree gets
a chance to shutdown gracefully still:

[09:29:37] INFO: Service mariadb exited with code 256 (by signal 9)
s6-rc: info: service mariadb-core successfully stopped
s6-rc: info: service mariadb-pre: stopping
...

The 18s timeout of S6_SERVICES_GRACETIME seems to apply for legacy
services only which are no longer used in this add-on. So drop this
definition.

Fixes #3566

* Increase shutdown timeout to 5 minutes