Increase mqtt max_queued_messages to 8192
The default of 1000 was not sufficent when the user
has more than 1000 entities and startup is so busy
that is cannot ready the buffer before it fills up
and the broker starts dropping messages.

While we have increased the mqtt receive buffer in
core, this only solves the problem on fast systems.