matter_server: Bump Python Matter server to 6.3.0
Update to the lastest Python Matter server. This makes sure that the OTA
Provider update binary is already part of the add-on so that updates can
be tested in the upcomming beta phase.

This also adds an option to set the Bluetooth adapter, which is useful
to test local commissioning.