Use Debian Bookworm as base
WiringPi fails compilation due to missing symbols - according to the deCONZ forum it is not required anymore, therefore I removed it.

Debian Buster is out of support since a few weeks now.
This should be updated to Bookworm.
The new release has no package for wmii anymore and manual compilation seems to be not possible as various required libraries are not available anymore (Not even as source).
I've changed the window manager to dwm for now.

Additionally the netcat package has been renamed to netcat-traditional.