matter_server: Add optional settings for development and testing (#3744)
* matter_server: Add support for extra Matter Server arguments

This adds an optional configuration to pass additional command line
arguments to the Matter Server. This is useful for not commonly used
settings (e.g. for testing and debugging) which we don't want to expose
as actual add-on configuration.

* matter_server: Allow to set a custom version to be installed

Add an option to install a custom Matter Server and Matter SDK (CHIP
wheels) version. This is an extension to the existing beta flag, but
instead of simply allowing the latest version, a version can be
specified. This can be helpful for development and testing.

* Apply suggestions from code review

Co-authored-by: Martin Hjelmare <marhje52@gmail.com>

---------

Co-authored-by: Martin Hjelmare <marhje52@gmail.com>