Lint all script files in the repository
The "lint" workflow that runs for all pull requests in this repository
used the `ludeeus/action-shellcheck@2.0.0` action as a shellcheck
wrapper. However, that wrapper only lints files with known extensions
and known shebangs, leading to lots of of scripts in this repository not
being subject to linting even though they are technically bash scripts.

Furthermore, there was no visibility into this linting problem, because
the linting step neither printed the list of files processed nor was
there a way to run the linter script locally.

This change:

1. Replaces the usage of the `ludeeus/action-shellcheck@2.0.0` action
   with an inline script that is runnable locally on both Linux and macOS;

2. Ensures that all script files in the repository are subject to linting
   regardless of their shebang;

3. Uses `shellcheck` that comes preinstalled in GitHub Actions runners
   rather than downloading it every time;

4. Reformats shellcheck output with Actions "workflow commands" syntax,
   which should lead to nicer formatting of linting errors in GitHub
   pull requests view.