Fix/Refactor git-pull addon
There is a bug present causing git-pull to no longer be able to restart or validate the configuration after updating the git repo.

I suspect part of this is related bashio, possibly related to not using with-contenv.

In this commit/pr I'm making use of the addon support present in bashio to avoid direct use of jq, and api calls.

#3714