openthread_border_router: Try restore the RCP communication on error
Try to restore the communication with the RCP in case it stalls out.
Some users see "radio tx timeout" errors. It is not quite clear what
causes those, it could be firmware crashes or failures in the underlying
communication layer. With this option set, the OTBR will automatically
restore the communication, and if the fimrware supports it, also request
a crash dump (requires debug logging).