openthread_border_router: Recover from RCP communication issue (#3887)
* openthread_border_router: add project specific OpenThread config

Add a project specific OpenThread compile time configuration header
file for the OTBR add-on. This allows to configure OpenThread compile
time configs which are not exposed through the CMake build system.
This allows to drop the default route metric patch file and use the
config header file instead.

* openthread_border_router: Increase number of frag entries

Increase the number of mesh header fragmentation tag entries to
potentially lower the amount of "Failed to get forwarded frame priority"
notice messages in logs.

* openthread_border_router: Try restore the RCP communication on error

Try to restore the communication with the RCP in case it stalls out.
Some users see "radio tx timeout" errors. It is not quite clear what
causes those, it could be firmware crashes or failures in the underlying
communication layer. With this option set, the OTBR will automatically
restore the communication, and if the fimrware supports it, also request
a crash dump (requires debug logging).

* openthread_border_router: Bump version and update changelog

* Update openthread_border_router/CHANGELOG.md

* Update openthread_border_router/CHANGELOG.md