samba: Bind on enabled interfaces only
Do not try to bind on disabled interfaces. Those don't have an IP
address and Samba won't be able to bind them. Also, Samba does not
automatically bind to interfaces if they get enabled at runtime. So
simply only pass the ones which are enabled.

This avoid spurious DNS requests from Samba. it seems that if an
interface is not up, Samba tries to resolve the string (probably in an
attempt to find a useful address for the given string).