Drop build for deprecated archs, use latest native runners (#4251)
In #4248 breaking bump of the builder action was merged. This means that future
attempts at builds of deprecated architectures will fail. This change removes
the deprecated architectures from the build matrix and changes the build job to
use native runner for aarch64 build, while updating to latest runner versions
as well.